back/requirements.txt
fastapi == 0.65.1
uvicorn == 0.14.0
motor == 2.4.0
openai
python-dotenv==1.0.1
fitz
pymupdf